+++ tests/framework/helpers/FileHelperTest.php
@@ -1308,6 +1308,18 @@ class FileHelperTest extends TestCase
                     'pjpeg',
                 ],
             ],
+            [
+                'application/x-rar-compressed',
+                [
+                    'rar',
+                ],
+            ],
+            [
+                'application/x-rar',
+                [
+                    'rar',
+                ],
+            ],
         ];
     }
 
@@ -1330,6 +1342,8 @@ class FileHelperTest extends TestCase
             ['application/json', false, 'json'],
             ['image/jpeg', true, 'jpg'],
             ['image/jpeg', false, 'jpeg'],
+            ['application/x-rar-compressed', false, 'rar'],
+            ['application/x-rar', false, 'rar'],
         ];
     }
 }